refactor: move string manipulation functions to utils module
